feat: 인증 및 인가 절차 추가

+++ candy/marker/views.py
@@ -10,12 +10,15 @@ def marker_view(request):
         return render(request, 'marker/marker.html', {})
     # 지도 클릭 (마커 리스트 + 기본 페이지)
     if request.method == 'POST':
+        # 마커 안 눌렀을 때 예외처리
+        if not request.POST.get('latitude'):
+            return render(request, 'marker/marker.html')
         status = int(request.POST.get('status', 1))
         latitude = float(request.POST.get('latitude', 33.450701))
         longitude = float(request.POST.get('longitude', 126.570667))
         filtered_markers = Marker.objects.filter(
-            latitude__range=(latitude - 0.001, latitude + 0.001),
-            longitude__range=(longitude - 0.001, longitude + 0.001)
+            latitude__range=(latitude - 0.002, latitude + 0.002),
+            longitude__range=(longitude - 0.002, longitude + 0.002)
         )
         return render(request, 'marker/marker.html', {"markers": filtered_markers, "status": status})
 
@@ -24,6 +27,8 @@ def marker_detail_view(request, pk):
     marker = Marker.objects.get(pk=pk)
     # 댓글 작성
     if request.method == 'POST':
+        if not request.user.is_authenticated:
+            return redirect('marker:marker_detail', pk=pk)
         user = request.user
         body = request.POST.get('body')
         Comment.objects.create(
@@ -39,6 +44,9 @@ def marker_detail_view(request, pk):
 
 # 지도 작성 (미완성)
 def marker_edit_view(request):
+    # 미인증 유저 예외처리
+    if not request.user.is_authenticated:
+        return redirect('marker:markers')
     # 기본 페이지
     if request.method == 'GET':
         lat = request.GET.get('lat')
@@ -92,6 +100,8 @@ def marker_detail_delete(request, pk):
     # marker_delete = Marker.objects.get(pk=pk)
     try:
         marker_delete = Marker.objects.get(pk=pk)
+        if not request.user.pk == marker_delete.user.pk:
+            redirect('marker:marker_detail', pk=pk)
         marker_delete.delete()
     except Marker.DoesNotExist:
         marker_delete = None
@@ -101,6 +111,8 @@ def marker_detail_delete(request, pk):
 def marker_detail_update_page(request, pk):
     try:
         marker_update = get_object_or_404(Marker, pk=pk)
+        if not request.user.pk == marker_update.user.pk:
+            redirect('marker:marker_detail', pk=pk)
     except Marker.DoesNotExist:
         marker_update=None
     return render(request,'marker/update.html', {'marker':marker_update})
